pass the information client
