+++ src/pages/ReservationConfirm.js
@@ -1,7 +1,8 @@
 import { ReservationConfirmPayment } from '../components/Reservation/ReservationConfirmPayment'
-import { useSelector } from 'react-redux'
+import { useSelector, useDispatch } from 'react-redux'
 import axios from 'axios'
 import { useEffect } from 'react'
+import { getClient } from '../store/actions/Client.actions'
 
 const handler = window.ePayco.checkout.configure({
   key: process.env.REACT_APP_EPAYCO_PUBLIC_KEY,
@@ -39,6 +40,8 @@ function Response({ location }) {
 
 function ReservationConfirm() {
 
+  const dispatch = useDispatch()
+
   const today = new Date()
   const month = today.toLocaleString('es-CO', { month: 'long'}) 
   const date = `${ today.getDate() } de ${ month } de ${ today.getFullYear()}`
@@ -50,6 +53,23 @@ function ReservationConfirm() {
     }}) => {
       return { ...state} 
     })
+
+  useEffect(() => {
+    dispatch(getClient())
+  }, [])
+
+  const {data} = useSelector(
+    ({ clientReducer: {
+      data
+    }}) => {
+      return {
+        data
+      }
+  })
+
+  const clientData = data
+
+  const invoice = localStorage.getItem('reservation');
   
   const handleClick = () => {
     handler.open(
@@ -57,7 +77,7 @@ function ReservationConfirm() {
         //Parametros compra (obligatorio)
         name: `${nameRestaurantReservation}`,
         description: "Pago Reserva alamesa",
-        invoice: "#RESERVA ",
+        invoice: `${invoice}`,
         currency: "cop",
         amount: `${deposit}`,
         tax_base: "0",
@@ -74,15 +94,14 @@ function ReservationConfirm() {
         response: `${process.env.REACT_APP_BASE_URL}/response`,
 
         //Atributos cliente
-        name_billing: "Nombre del cliente",
-        address_billing: "dirección cliente",
+        name_billing: `${clientData.name} ${clientData.lastName}`,
+        address_billing: `${clientData.address}`,
         type_doc_billing: "cc",
-        mobilephone_billing: "1098971546",
-        number_doc_billing: "3104567891",
+        mobilephone_billing: `${clientData.phone}`,
+        number_doc_billing: `${clientData.identification}`,
 
        //atributo deshabilitación metodo de pago
         methodsDisable: ["SP","CASH"]
-
       }
     )
   }
